refactor: rename node and update config
